requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apify_client==2.2.1
apify_shared==2.1.0
beautifulsoup4==4.14.2
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
impit==0.8.0
jiter==0.12.0
lxml==6.0.2
more-itertools==10.8.0
openai==2.7.2
playwright==1.55.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
